Use the correct variable types.

+++ app/src/main/java/org/tvheadend/tvhclient/utils/MenuUtils.java
@@ -253,7 +253,7 @@ public class MenuUtils {
         activity.startService(intent);
     }
 
-    public void handleMenuPlaySelection(long channelId, long dvrId) {
+    public void handleMenuPlaySelection(int channelId, int dvrId) {
         Activity activity = this.activity.get();
         if (activity == null) {
             return;
